requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
blinker==1.6.2
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
Flask==2.3.2
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.18.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
gspread==5.9.0
gspread-dataframe==3.3.0
httplib2==0.22.0
idna==3.4
ipykernel==6.23.0
ipython==8.13.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.2.0
jupyter_core==5.3.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.2.2
packaging==23.1
pandas==2.0.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.0
prompt-toolkit==3.0.38
protobuf==4.23.0
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyDrive==1.3.1
Pygments==2.15.1
pyngrok==6.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
stack-data==0.6.2
tornado==6.3.1
traitlets==5.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.3.6
